refactor: allocate send and recv buffers once
Rename traceroute_response_t to traceroute_recv_packet_t

Create the struct traceroute_send_packet_s to store the send buffer

Allocate the buffers in init_traceroute_conf

Add a function to free traceroute config

The choice has been made not to free the traceroute config if set_sockopt_ttl fails
This is to avoid making the code too long.
If this happens, the OS will automatically free leaks that are still reachable.